refactor: modularize receive view
