+++ src/app/(app)/(layout)/wallet/receive/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Receive } from '@/views/wallet/Receive';
+import { Receive } from '@/views/wallet/receive/Receive';
 
 export default function Page() {
   return <Receive />;
